OK Going to mothball this now Main bug needs fixing is the Adverts are not showing in Editor mode. Not investigated this
Unfinished job is downloading Shop items as addressables. Yes it works for buying tracks, but there's no "Buy A Track" button. You can still buy a track buy going into the shop for a scenery item and switching to the Track tab.

It works using the ShopItemDownloader which is called from the ShopPenel.

The purchasable tracks are still hard coded in the Shop.cs so we need a way of loading addressable ShopItems as well as the actual Purchasable assets.
See the shopItems have a picture and a price.

Downloading tracks is a special case because they are text assets. Also need to write code that downloads prefabs

Downloaded tracks could be deleted but they are still marked as purchased in Config.dat